Fixed Door Placement
Fixed the placement of doors by making it use the size of the sprite for placement.

Also started to create the script that allows the player to move through doors and from room to room.

+++ Assets/Scripts/RoomGeneration.cs
@@ -7,7 +7,7 @@ public class RoomGeneration : MonoBehaviour {
     public GameObject room, roomHolder;
     public int maxRooms;
     public int roomsCreated;
-    public int roomHeight, roomWidth;
+    public float roomHeight, roomWidth;
     public List<Vector2> placedRoomsCoords;
     Vector2 spawnCoordinates;
 
@@ -15,12 +15,12 @@ public class RoomGeneration : MonoBehaviour {
     public bool fullyGenerated;
 
 	void Start ()
-    {
+    { 
         fullyGenerated = false;
         spawnCoordinates = transform.position;
         InstantiateRoom();
         canGenerate = true;
-	}
+    }
 
     public void GenerateRooms()
     {
